add card

+++ src/main/kotlin/com/example/techcardservice/service/CategoryService.kt
@@ -2,6 +2,7 @@ package com.example.techcardservice.service
 
 import com.example.techcardservice.dto.CategoryDto
 import com.example.techcardservice.repository.CategoryRepository
+import com.example.techcardservice.repository.ComponentRepository
 import com.example.techcardservice.repository.mapper.CardMapper
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
@@ -9,11 +10,15 @@ import org.springframework.stereotype.Service
 import java.util.logging.Logger
 
 @Service
-class CategoryService(val categoryRepository: CategoryRepository,
-    val cardMapper: CardMapper) {
+class CategoryService(
+    val categoryRepository: CategoryRepository,
+    val componentRepository: ComponentRepository,
+    val cardMapper: CardMapper
+) {
     companion object {
         val logger = Logger.getLogger(CategoryService::class.java.name)
     }
+
     fun getCategory(pageable: Pageable): Page<CategoryDto> {
         logger.info("get all category")
         return categoryRepository.findAll(pageable).map { categoryEntity ->
@@ -29,9 +34,14 @@ class CategoryService(val categoryRepository: CategoryRepository,
         return categoryRepository.save(cardMapper.toCategoryEntity(categoryDto)).id
     }
 
-    fun deleteCategory(id: Long) {
+    fun deleteCategory(id: Long): Boolean {
         logger.info("delete category with id $id")
-        categoryRepository.deleteById(id)
+        val byCategory = componentRepository.getByCategoryId(id)
+        if (byCategory.isEmpty()) {
+            categoryRepository.deleteById(id)
+            return true
+        }
+        return false
     }
 
 }
